requirements.txt
django
mysqlclient
mysql-connector
mysql-connector-python